Siuuum
The agent spawn randomly in a circle of radius 1 near the origin, this allows it  to not overfit on a single action.  
Added multiple episode per rollout, this allows to not overfit on a single strategy.
Changed the overall stategy
Added the possibility to choose the initial position of agents and landmark during the env.reset().